examples/sandbox/vision-poc/requirements.txt
openai>=1.76
pydantic>=2.11.5
python-dotenv>=1.0.0
sqlalchemy>=2.0.41
